= Refactor @Serverconfig to valid relative path swapped on build
