+++ backpack.config.js
@@ -5,8 +5,8 @@ module.exports = {
     config.entry.main = './server/index.js';
     config.externals = [config.externals];
     config.externals.push((ctx, request, callback) => {
-      if (/^@ServerConfig\//.test(request)) {
-        const res = request.replace('@ServerConfig', '../server/config');
+      if (/^\.\.\/config/.test(request)) {
+        const res = request.replace('../config', '../server/config');
         return callback(null, res);
       }
       return callback();
