+++ server/util/gcloudPub.js
@@ -6,7 +6,7 @@ import {
 const PubSub = require('@google-cloud/pubsub');
 const uuidv4 = require('uuid/v4');
 
-const config = require('@ServerConfig/config.js'); // eslint-disable-line import/no-extraneous-dependencies
+const config = require('../config/config.js'); // eslint-disable-line import/no-extraneous-dependencies
 
 const pubsub = new PubSub();
 const topics = [
