+++ server/util/jwt.js
@@ -4,7 +4,7 @@ const crypto = require('crypto');
 const jwt = require('jsonwebtoken');
 const expressjwt = require('express-jwt');
 const { UnauthorizedError } = require('express-jwt');
-const config = require('@ServerConfig/config.js'); // eslint-disable-line import/no-extraneous-dependencies
+const config = require('../config/config.js'); // eslint-disable-line import/no-extraneous-dependencies
 
 let secret = config.JWT_SECRET;
 if (!secret) {
