+++ test/api/payment.test.js
@@ -11,7 +11,7 @@ import {
 const jwt = require('jsonwebtoken');
 const axiosist = require('axiosist');
 
-const app = require('../../build/main.js'); // eslint-disable-line import/no-unresolved
+const app = require('../../server/index.js'); // eslint-disable-line import/no-unresolved
 
 
 test('PAYMENT: Payment. Case: Invalid address.', async (t) => {
